+++ packages/app/src/test/repo-link.test.tsx
@@ -67,6 +67,7 @@ function makeClient() {
     resolveFriction: () => Promise.resolve(),
     addressFriction: () => Promise.resolve(),
     reopenFriction: () => Promise.resolve(),
+    captureViewport: () => Promise.resolve(new Uint8Array()),
     revealInFinder: () => Promise.resolve(),
     listFindings: () => Promise.resolve([]),
     signalFinding: () => Promise.resolve(),